+++ src/main/java/gregtech/common/render/GT_CapeRenderer.java
@@ -41,7 +41,7 @@ public class GT_CapeRenderer
         if (aPlayer.isInvisible()) {
             return;
         }
-        if (GT_Utility.getPotion(aPlayer, Integer.valueOf(Potion.invisibility.id).intValue())) {
+        if (GT_Utility.getPotion(aPlayer, Integer.valueOf(Potion.invisibility.id))) {
             return;
         }
         try {
